Seperate and test MessageTimeout
